move: SecondNavBar to actionsAndAttributes
