+++ apps/web/app/(app)/environments/[environmentId]/(actionsAndAttributes)/SecondNavBar.tsx
